requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
emoji==2.14.1
idna==3.10
pillow==11.1.0
requests==2.32.3
ttkbootstrap==1.10.1
urllib3==2.3.0
